move table into Board component
