+++ src/components/index.js
@@ -1,0 +1,5 @@
+import Keyboard from './Keyboard/Keyboard';
+import Board from './Board/Board';
+
+
+export { Keyboard, Board };
